notebook-server/requirements.txt
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
docker==4.2.2
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Classy==0.6.10
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
psutil==5.7.2
pycparser==2.20
PyMySQL==0.9.3
PyYAML==5.3
requests==2.24.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.10
visitor==0.1.3
websocket-client==0.57.0
Werkzeug==1.0.0
